Change the GIF default loopcount to 1
React Native iOS has a bug that transforms a meta loopCount of 1 to zero, thus leading to infinitely looping gif instead of 1 time loop. On android there is no such problem. This is temporarily circumvented by setting the looping GIFs to have a large loop value (> 65000) and making this modification here.